Remove IMessageLock
Remove the IMessageLock interface, just leaving IMessageLockAsync.
Resolves #458.
